Prove some things about u8s in their own file
I want to be able to be able to reason about getting and setting individual
bits, which is not so easy with the LOG___ functions built in. I am hoping
that the strategy I used in here will make things easier. It did admit the
bit manipulation I wanted so far.